pet-feeder-iot-main/requirements.txt
tensorflow==2.18.0
firebase-admin==6.6.0
opencv-python==4.11.0.86
Flask==3.1.0
pyngrok==7.2.3
python-vlc==3.0.21203